refactor: AffiliationContainer 컴포넌트 변경
