+++ src/components/page/student/DepartmentInfo/DepartmentInfoApplyLocker/index.module.scss
@@ -1,3 +1,7 @@
+.skeleton {
+  height: 30rem;
+}
+
 .container {
   @include flexSort(column, null, null, 1.5rem);
 }
