+++ src/components/page/student/DepartmentInfo/index.module.scss
@@ -1,8 +1,8 @@
-.skeleton {
-  margin: 2rem 3rem 0 3rem;
-  height: 5rem;
+.container {
+  padding: 2rem 3rem;
+  @include flexSort(column, null, null, 3rem);
 }
 
-.container {
+.contentContainer {
   @include flexSort(column, null, null, 9rem);
 }
